requirements.txt
bson==0.5.10
scikit-learn==1.3.0
pymongo==4.4.1
scipy==1.11.1
ncls==0.0.68
nltk==3.8.1
numpy==1.25.1
lemminflect==0.2.3
wordfreq==3.0.3
werkzeug==2.3.6
spacy==3.6.0
PyYAML==5.4.1
pyphen==0.14.0
pyspotlight==0.7.2
jpype1==1.4.1
